+++ backend/infrahub/core/branch/tasks.py
@@ -1,9 +1,8 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Sequence
+from typing import TYPE_CHECKING, Sequence
 from uuid import uuid4
 
-import pydantic
 from prefect import flow, get_run_logger
 from prefect.client.schemas.objects import State  # noqa: TC002
 from prefect.states import Completed, Failed
@@ -12,6 +11,7 @@ from infrahub import config, lock
 from infrahub.context import InfrahubContext  # noqa: TC001  needed for prefect flow
 from infrahub.core import registry
 from infrahub.core.branch import Branch
+from infrahub.core.branch.creator import BranchCreator
 from infrahub.core.branch.enums import BranchStatus
 from infrahub.core.changelog.diff import DiffChangelogCollector, MigrationTracker
 from infrahub.core.constants import DiffAction, MutationAction
@@ -34,7 +34,6 @@ from infrahub.core.validators.models.validate_migration import SchemaValidateMig
 from infrahub.core.validators.tasks import schema_validate_migrations
 from infrahub.dependencies.registry import get_component_registry
 from infrahub.events.branch_action import (
-    BranchCreatedEvent,
     BranchDeletedEvent,
     BranchMergedEvent,
     BranchMigratedEvent,
@@ -42,7 +41,7 @@ from infrahub.events.branch_action import (
 )
 from infrahub.events.models import EventMeta, InfrahubEvent
 from infrahub.events.node_action import get_node_event
-from infrahub.exceptions import BranchNotFoundError, ValidationError
+from infrahub.exceptions import ValidationError
 from infrahub.generators.constants import GeneratorDefinitionRunSource
 from infrahub.graphql.mutations.models import BranchCreateModel  # noqa: TC001
 from infrahub.workers.dependencies import get_component, get_database, get_event_service, get_workflow
@@ -51,7 +50,6 @@ from infrahub.workflows.catalogue import (
     BRANCH_MERGE_POST_PROCESS,
     DIFF_REFRESH_ALL,
     DIFF_UPDATE,
-    GIT_REPOSITORIES_CREATE_BRANCH,
     IPAM_RECONCILIATION,
     TRIGGER_ARTIFACT_DEFINITION_GENERATE,
     TRIGGER_GENERATOR_DEFINITION_RUN,
@@ -63,10 +61,6 @@ if TYPE_CHECKING:
 
     from infrahub.core.changelog.models import NodeChangelog
     from infrahub.database import InfrahubDatabase
-    from infrahub.lock import InfrahubLockRegistry
-    from infrahub.services.adapters.event import InfrahubEventService
-    from infrahub.services.adapters.workflow import InfrahubWorkflow
-    from infrahub.services.component import InfrahubComponent
 
 
 @flow(name="branch-migrate", flow_run_name="Apply migrations to branch {branch}")
@@ -483,74 +477,6 @@ async def validate_branch(branch: str) -> State:
         if has_conflicts:
             return Failed(message="branch has some conflicts")
         return Completed(message="branch is valid")
-
-
-class BranchCreator:
-    def __init__(
-        self,
-        db: InfrahubDatabase,
-        lock_registry: InfrahubLockRegistry,
-        component: InfrahubComponent,
-        event_service: InfrahubEventService,
-        workflow: InfrahubWorkflow,
-    ) -> None:
-        self.db = db
-        self.lock_registry = lock_registry
-        self.component = component
-        self.event_service = event_service
-        self.workflow = workflow
-
-    async def create(self, model: BranchCreateModel, context: InfrahubContext) -> None:
-        try:
-            await Branch.get_by_name(db=self.db, name=model.name)
-            raise ValidationError(f"The branch {model.name} already exists")
-        except BranchNotFoundError:
-            pass
-
-        data_dict: dict[str, Any] = dict(model)
-        data_dict.pop("is_isolated", None)
-
-        try:
-            obj = Branch(**data_dict)
-        except pydantic.ValidationError as exc:
-            error_msgs = [f"invalid field {error['loc'][0]}: {error['msg']}" for error in exc.errors()]
-            raise ValidationError("\n".join(error_msgs)) from exc
-
-        # distributed lock to prevent creating multiple branches with the same name
-        async with self.lock_registry.get(name=model.name, namespace="branch_create", local=False):
-            # Re-check existence under the lock to prevent TOCTOU race
-            try:
-                await Branch.get_by_name(db=self.db, name=model.name)
-                raise ValidationError(f"The branch {model.name} already exists")
-            except BranchNotFoundError:
-                pass
-
-            async with self.lock_registry.local_schema_lock():
-                # Copy the schema from the origin branch and set the hash and the schema_changed_at value
-                origin_schema = registry.schema.get_schema_branch(name=obj.origin_branch)
-                new_schema = origin_schema.duplicate(name=obj.name)
-                registry.schema.set_schema_branch(name=obj.name, schema=new_schema)
-                obj.update_schema_hash()
-                await obj.save(db=self.db, user_id=context.account.account_id)
-
-                # Add Branch to registry
-                registry.branch[obj.name] = obj
-                await self.component.refresh_schema_hash(branches=[obj.name])
-
-        event = BranchCreatedEvent(
-            branch_name=obj.name,
-            branch_id=str(obj.uuid),
-            sync_with_git=obj.sync_with_git,
-            meta=EventMeta.from_context(context=context, branch=registry.get_global_branch()),
-        )
-        await self.event_service.send(event=event)
-
-        if obj.sync_with_git:
-            await self.workflow.submit_workflow(
-                workflow=GIT_REPOSITORIES_CREATE_BRANCH,
-                context=context,
-                parameters={"branch": obj.name, "branch_id": str(obj.uuid)},
-            )
 
 
 @flow(name="create-branch", flow_run_name="Create branch {model.name}")
